+++ pkg/text/interface.go
@@ -1,6 +1,7 @@
 package text
 
 import (
+	"context"
 	"html/template"
 
 	"github.com/zostay/today/pkg/ref"
@@ -9,14 +10,14 @@ import (
 // Resolver is the interface used to retrieve text for a scripture passage.
 type Resolver interface {
 	// Verse turns a reference into a string of text.
-	Verse(ref *ref.Resolved) (string, error)
+	Verse(ctx context.Context, ref *ref.Resolved) (string, error)
 
 	// VerseHTML turns a reference into a string of HTML.
-	VerseHTML(ref *ref.Resolved) (template.HTML, error)
+	VerseHTML(ctx context.Context, ref *ref.Resolved) (template.HTML, error)
 
 	// VersionInformation returns the metadata for the version of the Bible
 	// used for the verse.
-	VersionInformation() (*Version, error)
+	VersionInformation(ctx context.Context) (*Version, error)
 }
 
 // Version is the metadata for the version of the Bible used for the verse.